refactor(api): file structure
