+++ api/src/routes/mod.rs
@@ -1,0 +1,2 @@
+mod get_self;
+pub(crate) use get_self::get_self;
